requirements.txt
fastapi
uvicorn
sqlalchemy
langchain==0.3.19
openai
requests
python-dotenv
pypdf2
pandas
pillow
pydantic
faker
langchain-openai==0.3.18
langchain-community==0.3.18
langchain-text-splitters
langchain_core
langchain-perplexity==0.1.1
python-multipart
langchain-groq==0.3.2
fastapi-mcp